Fix QNN runner KV cache bitwidth in Android JNI
The QNN runner was hardcoded to use Runner<uint16_t>, but all current
Llama quantization recipes use annotate_kv_8bit for 8-bit KV cache.
This mismatch caused the KV cache data to be misinterpreted, resulting
in degenerate output (repetitive real words like "Nigeria Nigeria...")
while the model otherwise ran correctly on the HTP NPU.

Authored with Claude

+++ extension/android/jni/jni_layer_llama.cpp
@@ -216,7 +216,7 @@ class ExecuTorchLlmJni : public facebook::jni::HybridClass<ExecuTorchLlmJni> {
           data_files_vector,
           executorch::extension::Module::LoadMode::MmapUseMlockIgnoreErrors);
       std::string decoder_model = "llama3"; // use llama3 for now
-      runner_ = std::make_unique<example::Runner<uint16_t>>( // QNN runner
+      runner_ = std::make_unique<example::Runner<uint8_t>>( // QNN runner (8-bit KV cache)
           std::move(module),
           decoder_model.c_str(),
           model_path->toStdString().c_str(),
